requirements.txt
mlflow-skinny==2.17.2
xgboost==2.1.3
scikit-learn==1.5.2
shap==0.46.0
python-dateutil==2.9.0
pandas==2.2.3
numpy==2.2.1
